Downloads/wallet_api/wallet_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
requests
python-dotenv